allow apos.doc.insert in init of any module initialized after the db module. Divorces uploadfs from the attachment module, it is now initialized by its own module and made available as apos.uploadfs.
