+++ defaults.js
@@ -4,7 +4,7 @@ module.exports = {
     '@apostrophecms/util': {},
     '@apostrophecms/task': {},
     '@apostrophecms/schema': {},
-    '@apostrophecms/attachment': {},
+    '@apostrophecms/uploadfs': {},
     '@apostrophecms/asset': {},
     '@apostrophecms/launder': {},
     '@apostrophecms/http': {},
@@ -13,6 +13,7 @@ module.exports = {
     '@apostrophecms/lock': {},
     '@apostrophecms/cache': {},
     '@apostrophecms/migration': {},
+    '@apostrophecms/attachment': {},
     '@apostrophecms/express': {},
     '@apostrophecms/url': {},
     '@apostrophecms/template': {},
