+++ modules/@apostrophecms/attachment/index.js
@@ -2,44 +2,13 @@ const _ = require('lodash');
 const path = require('path');
 const fs = require('fs');
 const Promise = require('bluebird');
-const uploadfs = require('uploadfs');
-const mkdirp = require('mkdirp');
 
 module.exports = {
   options: { alias: 'attachment' },
-  cascades: [ 'imageSizes' ],
-  imageSizes: {
-    add: {
-      max: {
-        width: 1600,
-        height: 1600
-      },
-      full: {
-        width: 1140,
-        height: 1140
-      },
-      'two-thirds': {
-        width: 760,
-        height: 760
-      },
-      'one-half': {
-        width: 570,
-        height: 700
-      },
-      'one-third': {
-        width: 380,
-        height: 700
-      },
-      'one-sixth': {
-        width: 190,
-        height: 350
-      }
-    }
-  },
 
   async init(self) {
     self.name = 'attachment';
-
+    self.imageSizes = self.apos.modules['@apostrophecms/uploadfs'].imageSizes;
     self.fileGroups = self.options.fileGroups || [
       {
         name: 'images',
@@ -95,40 +64,14 @@ module.exports = {
 
     self.sizeAvailableInTrash = self.options.sizeAvailableInTrash || 'one-sixth';
 
-    // uploadfs expects an array
-    self.imageSizes = Object.keys(self.imageSizes).map(name => ({
-      name,
-      ...self.imageSizes[name]
-    }));
-
-    const uploadfsDefaultSettings = {
-      backend: 'local',
-      uploadsPath: self.apos.rootDir + '/public/uploads',
-      uploadsUrl: (self.apos.baseUrl || '') + self.apos.prefix + '/uploads',
-      tempPath: self.apos.rootDir + '/data/temp/uploadfs',
-      imageSizes: self.imageSizes
-    };
-
-    self.uploadfsSettings = {};
-    _.merge(self.uploadfsSettings, uploadfsDefaultSettings);
-
-    _.merge(self.uploadfsSettings, self.options.uploadfs || {});
-
-    if (process.env.APOS_S3_BUCKET) {
-      _.merge(self.uploadfsSettings, {
-        backend: 's3',
-        endpoint: process.env.APOS_S3_ENDPOINT,
-        secret: process.env.APOS_S3_SECRET,
-        key: process.env.APOS_S3_KEY,
-        bucket: process.env.APOS_S3_BUCKET,
-        region: process.env.APOS_S3_REGION
-      });
-    }
-
     self.rescaleTask = require('./lib/tasks/rescale.js')(self);
-    await self.initUploadfs();
     self.addFieldType();
     self.enableBrowserData();
+
+    self.db = await self.apos.db.collection('aposAttachments');
+    await self.db.createIndex({ docIds: 1 });
+    await self.db.createIndex({ trashDocIds: 1 });
+    self.addFixLengthPropertyMigration();
   },
 
   tasks(self) {
@@ -212,22 +155,9 @@ module.exports = {
   },
   handlers(self) {
     return {
-      'apostrophe:modulesReady': {
-        // Delay setting up the collection until other modules
-        // are ready in a normal startup. Necessary because uploadfs must
-        // be available during asset builds but the database must not be
-        async enableCollection() {
-          self.db = await self.apos.db.collection('aposAttachments');
-          await self.db.createIndex({ docIds: 1 });
-          await self.db.createIndex({ trashDocIds: 1 });
-        },
-        addFixLengthPropertyMigration() {
-          self.addFixLengthPropertyMigration();
-        }
-      },
       'apostrophe:destroy': {
         async destroyUploadfs() {
-          await Promise.promisify(self.uploadfs.destroy)();
+          await Promise.promisify(self.apos.uploadfs.destroy)();
         }
       },
       '@apostrophecms/doc-type:afterSave': {
@@ -256,22 +186,6 @@ module.exports = {
   },
   methods(self) {
     return {
-      async initUploadfs() {
-        safeMkdirp(self.uploadfsSettings.uploadsPath);
-        safeMkdirp(self.uploadfsSettings.tempPath);
-        self.uploadfs = uploadfs();
-        await Promise.promisify(self.uploadfs.init)(self.uploadfsSettings);
-        function safeMkdirp(path) {
-          try {
-            mkdirp.sync(path);
-          } catch (e) {
-            if (!require('fs').existsSync(path)) {
-              throw e;
-            }
-          }
-        }
-      },
-
       addFieldType() {
         self.apos.schema.addFieldType({
           name: self.name,
@@ -452,7 +366,7 @@ module.exports = {
         info.md5 = await self.apos.util.md5File(file.path);
         if (self.isSized(extension)) {
           // For images we correct automatically for common file extension mistakes
-          const result = await Promise.promisify(self.uploadfs.copyImageIn)(file.path, '/attachments/' + info._id + '-' + info.name);
+          const result = await Promise.promisify(self.apos.uploadfs.copyImageIn)(file.path, '/attachments/' + info._id + '-' + info.name);
           info.extension = result.extension;
           info.width = result.width;
           info.height = result.height;
@@ -465,7 +379,7 @@ module.exports = {
           // For non-image files we have to trust the file extension
           // (but we only serve it as that content type, so this should
           // be reasonably safe)
-          await Promise.promisify(self.uploadfs.copyIn)(file.path, '/attachments/' + info._id + '-' + info.name + '.' + info.extension);
+          await Promise.promisify(self.apos.uploadfs.copyIn)(file.path, '/attachments/' + info._id + '-' + info.name + '.' + info.extension);
         }
         if (options.permissions !== false) {
           info.ownerId = req.user && req.user._id;
@@ -499,10 +413,10 @@ module.exports = {
         // Pull the original out of cloud storage to a temporary folder where
         // it can be cropped and popped back into uploadfs
         const originalFile = '/attachments/' + info._id + '-' + info.name + '.' + info.extension;
-        const tempFile = self.uploadfs.getTempPath() + '/' + self.apos.util.generateId() + '.' + info.extension;
+        const tempFile = self.apos.uploadfs.getTempPath() + '/' + self.apos.util.generateId() + '.' + info.extension;
         const croppedFile = '/attachments/' + info._id + '-' + info.name + '.' + crop.left + '.' + crop.top + '.' + crop.width + '.' + crop.height + '.' + info.extension;
-        await Promise.promisify(self.uploadfs.copyOut)(originalFile, tempFile);
-        await Promise.promisify(self.uploadfs.copyImageIn)(tempFile, croppedFile, { crop: crop });
+        await Promise.promisify(self.apos.uploadfs.copyOut)(originalFile, tempFile);
+        await Promise.promisify(self.apos.uploadfs.copyImageIn)(tempFile, croppedFile, { crop: crop });
         crops.push(crop);
         await self.db.updateOne({
           _id: info._id
@@ -549,7 +463,7 @@ module.exports = {
         }
         let path = '/attachments/' + attachment._id + '-' + attachment.name;
         if (!options.uploadfsPath) {
-          path = self.uploadfs.getUrl() + path;
+          path = self.apos.uploadfs.getUrl() + path;
         }
         // Attachments can have "one true crop," or a crop can be passed with the options.
         // For convenience, be tolerant if options.crop is passed but doesn't
@@ -995,7 +909,7 @@ module.exports = {
       // (usually `one-sixth`) remains available except when removing. This operation is carried
       // out across all sizes and crops.
       async alterAttachment(attachment, action) {
-        let method = self.uploadfs[action];
+        let method = self.apos.uploadfs[action];
         method = Promise.promisify(method);
         await original();
         await crops();
@@ -1065,17 +979,17 @@ module.exports = {
         }
       },
       async migrateToDisabledFileKeyTask(argv) {
-        await Promise.promisify(self.uploadfs.migrateToDisabledFileKey)();
+        await Promise.promisify(self.apos.uploadfs.migrateToDisabledFileKey)();
       },
       async migrateFromDisabledFileKeyTask(argv) {
-        await Promise.promisify(self.uploadfs.migrateFromDisabledFileKey)();
+        await Promise.promisify(self.apos.uploadfs.migrateFromDisabledFileKey)();
       },
       getBrowserData(req) {
         return {
           action: self.action,
           fileGroups: self.fileGroups,
           name: self.name,
-          uploadsUrl: self.uploadfs.getUrl(),
+          uploadsUrl: self.apos.uploadfs.getUrl(),
           croppable: self.croppable,
           sized: self.sized
         };
